move checkIfExerciseDueDateIsReached to SubmissionService and add check for programming exercises
